Reduce # of Bottles per Bucket in Crafting Recipe
Bottles normally hold 1000mB of fluid, same as a bucket.  There isn't really a reason to allow players to duplicate fluids in this way, and really it just doesn't make sense to milk with a bottle when you can just milk with a bucket and get 2 bottles out of it.