Repo Update
Honestly there's so much here since the last commit a few years ago, I'm not sure what the intention of each mod in here is so I'm just going to commit what I have currently.

There are a lot of config files left in the repo for mods I have since removed, which will need to be removed themselves eventually.